Move KeyboardWalkNew to KeyboardWalk purpose
