--- module/behaviour/strategy/KeyboardWalkNew/README.md
@@ -1,9 +1,0 @@
-# KeyboardWalk
-
-## Description
-
-## Usage
-
-## Emits
-
-## Dependencies
