+++ roles/webots/directorkeyboardwalk.role
@@ -29,7 +29,6 @@ nuclear_role(
   # Strategies
   strategy::FallRecovery
   strategy::StandStill
-  behaviour::strategy::KeyboardWalkNew
   # Purpose
   purpose::KeyboardWalk
 )
